+++ src/components/Lazy.js
@@ -5,16 +5,16 @@ import Observer from '@researchgate/react-intersection-observer'
 
 import { propsWithNoScriptRender } from '../lib/wrap'
 
-export default class Lazy extends React.PureComponent {
+class Lazy extends React.PureComponent {
     constructor(props) {
         super(props)
 
         this.state = { show: false }
 
-        this.onViewport = this.onViewport.bind(this)
+        this.handleViewport = this.handleViewport.bind(this)
     }
 
-    onViewport(event, unobserve) {
+    handleViewport(event, unobserve) {
         if (!event.isIntersecting || !this.props.visible) {
             return
         }
@@ -38,23 +38,25 @@ export default class Lazy extends React.PureComponent {
             clientOnly,
             component,
             cushion,
+            forwardedRef: ref,
             ltIE9,
             visible,
             onLoad,
             onViewport,
             threshold,
             viewport,
-            ...props
+            ...rest
         } = this.props
 
+        const props = Object.assign({ ref }, rest)
         const isClientRender = clientOnly || this.state.show
 
         return (
-            <Observer onChange={this.onViewport} root={viewport} rootMargin={cushion} threshold={threshold}>
+            <Observer onChange={this.handleViewport} root={viewport} rootMargin={cushion} threshold={threshold}>
                 {React.createElement(
                     component,
                     isClientRender ? props : propsWithNoScriptRender(children, ltIE9, props),
-                    isClientRender && this.state.show && visible ? children : null
+                    isClientRender && this.state.show && visible ? children : null,
                 )}
             </Observer>
         )
@@ -65,20 +67,24 @@ Lazy.defaultProps = {
     clientOnly: false,
     component: 'div',
     ltIE9: false,
-    visible: true
+    visible: true,
 }
 
 Lazy.propTypes = {
+    clientOnly: PropTypes.bool,
     children: PropTypes.node,
     component: PropTypes.any,
     cushion: PropTypes.string,
-    clientOnly: PropTypes.bool,
+    forwardedRef: PropTypes.any,
     ltIE9: PropTypes.bool,
     onLoad: PropTypes.func,
     onViewport: PropTypes.func,
     threshold: PropTypes.oneOfType([PropTypes.number, PropTypes.arrayOf(PropTypes.number)]),
     viewport: PropTypes.oneOfType(
-        [PropTypes.string].concat(typeof HTMLElement === 'undefined' ? [] : PropTypes.instanceOf(HTMLElement))
+        [PropTypes.string].concat(typeof HTMLElement === 'undefined' ? [] : PropTypes.instanceOf(HTMLElement)),
     ),
-    visible: PropTypes.bool
+    visible: PropTypes.bool,
 }
+
+// eslint-disable-next-line react/display-name,react/no-multi-comp
+export default React.forwardRef((props, ref) => <Lazy {...props} forwardedRef={ref} />)