+++ src/components/LazyGroup.js
@@ -6,18 +6,18 @@ import { countTypesTags, wrapTypesToLazyChild, wrapTypesToNoScript } from '../li
 import DefaultWrapper from './DefaultWrapper'
 import Observer from '@researchgate/react-intersection-observer'
 
-export default class LazyGroup extends React.PureComponent {
+class LazyGroup extends React.PureComponent {
     constructor(props) {
         super(props)
 
         this.loadedImgTags = 0
         this.state = { imgTagCount: null, loadedAt: null, viewportAt: null }
 
-        this.onImgLoaded = this.onImgLoaded.bind(this)
-        this.onViewport = this.onViewport.bind(this)
+        this.handleViewport = this.handleViewport.bind(this)
+        this.handleImgLoaded = this.handleImgLoaded.bind(this)
     }
 
-    onImgLoaded() {
+    handleImgLoaded() {
         this.loadedImgTags++
 
         if (this.loadedImgTags === this.state.imgTagCount) {
@@ -26,7 +26,7 @@ export default class LazyGroup extends React.PureComponent {
         }
     }
 
-    onViewport(event, unobserve) {
+    handleViewport(event, unobserve) {
         const { children, childrenToWrap, onLoad, onViewport, visible } = this.props
 
         if (!event.isIntersecting || !visible) {
@@ -48,7 +48,7 @@ export default class LazyGroup extends React.PureComponent {
         const viewportAt = Date.now()
         this.setState(
             { imgTagCount, loadedAt: !imgTagCount ? viewportAt : null, viewportAt },
-            !imgTagCount ? onLoad : null
+            !imgTagCount ? onLoad : null,
         )
     }
 
@@ -60,26 +60,27 @@ export default class LazyGroup extends React.PureComponent {
             clientOnly,
             component,
             cushion,
+            forwardedRef: ref,
             ltIE9,
             onLoad,
             onViewport,
             threshold,
             viewport,
             visible,
-            ...props
+            ...rest
         } = this.props
 
+        const props = Object.assign({ ref }, rest)
+
         return (
-            <Observer onChange={this.onViewport} root={viewport} rootMargin={cushion} threshold={threshold}>
+            <Observer onChange={this.handleViewport} root={viewport} rootMargin={cushion} threshold={threshold}>
                 {React.createElement(
                     component,
                     props,
                     // swap render once element is visible in viewport
                     clientOnly || this.state.viewportAt
-                        // replace elements with LazyChild
-                        ? wrapTypesToLazyChild(childrenToWrap, children, childWrapper, this.onImgLoaded)
-                        // wrap given element types to noscript and the given wrapper component
-                        : wrapTypesToNoScript(childrenToWrap, children, ltIE9, childWrapper)
+                        ? wrapTypesToLazyChild(childrenToWrap, children, childWrapper, this.handleImgLoaded, this.state.viewportAt != null)
+                        : wrapTypesToNoScript(childrenToWrap, children, ltIE9, childWrapper),
                 )}
             </Observer>
         )
@@ -92,22 +93,26 @@ LazyGroup.defaultProps = {
     clientOnly: false,
     component: 'div',
     ltIE9: false,
-    visible: true
+    visible: true,
 }
 
 LazyGroup.propTypes = {
+    clientOnly: PropTypes.bool,
     children: PropTypes.node,
-    childrenToWrap: PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.func, PropTypes.object, PropTypes.string])),
+    childrenToWrap: PropTypes.arrayOf(PropTypes.any),
     childWrapper: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
     component: PropTypes.any,
     cushion: PropTypes.string,
-    clientOnly: PropTypes.bool,
+    forwardedRef: PropTypes.any,
     ltIE9: PropTypes.bool,
     onLoad: PropTypes.func,
     onViewport: PropTypes.func,
     threshold: PropTypes.oneOfType([PropTypes.number, PropTypes.arrayOf(PropTypes.number)]),
     viewport: PropTypes.oneOfType(
-        [PropTypes.string].concat(typeof HTMLElement === 'undefined' ? [] : PropTypes.instanceOf(HTMLElement))
+        [PropTypes.string].concat(typeof HTMLElement === 'undefined' ? [] : PropTypes.instanceOf(HTMLElement)),
     ),
-    visible: PropTypes.bool
+    visible: PropTypes.bool,
 }
+
+// eslint-disable-next-line react/display-name,react/no-multi-comp
+export default React.forwardRef((props, ref) => <LazyGroup {...props} forwardedRef={ref} />)