+++ src/index.js
@@ -1,0 +1,2 @@
+export { default as LazyGroup } from './components/LazyGroup'
+export { default as Lazy } from './components/Lazy'