+++ src/index.jsx
@@ -1,2 +1,0 @@
-export LazyGroup from './components/LazyGroup'
-export Lazy from './components/Lazy'